Remove deleteaccount mutation
There doesn't seem to be any reason to allow users to delete their accounts.
It is likely that we'd never want to give a user this ability. Even if we might
offer an option like canceling one's account, soft-deletion seems like a more likely
choice. Removing this from the API to prevent confusion and possible issues that
might arise from deletion of an account.